tests: Move html func to test/util
