+++ button/test.js
@@ -1,6 +1,7 @@
 const tape = require('tape')
 const { qs } = require('qs')
 
+const { html } = require('../test/util')
 const components = require('..')
 components(require('@optoolco/tonic'))
 
@@ -247,14 +248,3 @@ tape('{{button-10}} has tabindex attribute', t => {
 
   t.end()
 })
-
-function html ([str, ...strings], ...values) {
-  let text = str
-  for (let i = 0; i < values.length; i++) {
-    text += values[i] + strings[i]
-  }
-
-  const tmpl = document.createElement('template')
-  tmpl.innerHTML = text
-  return tmpl.content.firstElementChild
-}
