+++ popover/test.js
@@ -1,3 +1,5 @@
+
+const { html } = require('../test/util')
 const components = require('..')
 components(require('@optoolco/tonic'))
 
@@ -36,14 +38,3 @@ popover.addEventListener('show', event => {
 })
 
 // TODO: write tests for popover.
-
-function html ([str, ...strings], ...values) {
-  let text = str
-  for (let i = 0; i < values.length; i++) {
-    text += values[i] + strings[i]
-  }
-
-  const tmpl = document.createElement('template')
-  tmpl.innerHTML = text
-  return tmpl.content.firstElementChild
-}
