+++ profile-image/test.js
@@ -1,3 +1,4 @@
+const { html } = require('../test/util')
 
 document.body.appendChild(html`
 <section id="profile-image">
@@ -111,14 +112,3 @@ document.body.appendChild(html`
 `)
 
 // TODO: write tests for profile-image
-
-function html ([str, ...strings], ...values) {
-  let text = str
-  for (let i = 0; i < values.length; i++) {
-    text += values[i] + strings[i]
-  }
-
-  const tmpl = document.createElement('template')
-  tmpl.innerHTML = text
-  return tmpl.content.firstElementChild
-}
