+++ select/test.js
@@ -1,3 +1,5 @@
+const { html } = require('../test/util')
+
 document.body.appendChild(html`
 <section id="select">
   <h2>Select</h2>
@@ -188,14 +190,3 @@ document.body.appendChild(html`
 `)
 
 // TODO: write tests
-
-function html ([str, ...strings], ...values) {
-  let text = str
-  for (let i = 0; i < values.length; i++) {
-    text += values[i] + strings[i]
-  }
-
-  const tmpl = document.createElement('template')
-  tmpl.innerHTML = text
-  return tmpl.content.firstElementChild
-}
