+++ tabs/test.js
@@ -2,6 +2,7 @@ const Tonic = require('@optoolco/tonic')
 const tape = require('tape')
 const { qs } = require('qs')
 
+const { html } = require('../test/util')
 const components = require('..')
 components(require('@optoolco/tonic'))
 
@@ -155,14 +156,3 @@ tape('{{tabs-3}} has correct default state', t => {
   t.ok(container, 'rendered')
   t.end()
 })
-
-function html ([str, ...strings], ...values) {
-  let text = str
-  for (let i = 0; i < values.length; i++) {
-    text += values[i] + strings[i]
-  }
-
-  const tmpl = document.createElement('template')
-  tmpl.innerHTML = text
-  return tmpl.content.firstElementChild
-}
