+++ toaster/test.js
@@ -1,6 +1,7 @@
 const tape = require('tape')
 const { qs } = require('qs')
 
+const { html } = require('../test/util')
 const components = require('..')
 components(require('@optoolco/tonic'))
 
@@ -274,14 +275,3 @@ tape('{{toaster}} is created on the right', async t => {
 
   t.end()
 })
-
-function html ([str, ...strings], ...values) {
-  let text = str
-  for (let i = 0; i < values.length; i++) {
-    text += values[i] + strings[i]
-  }
-
-  const tmpl = document.createElement('template')
-  tmpl.innerHTML = text
-  return tmpl.content.firstElementChild
-}
